* TA3D::slist is meant to be a small list, it is used to store the list of flying units in a map bloc, so it's unlikely it'll ever contain more than 32000 elements : its size member can be a short int NB: on big maps this saves a few MB OO!

+++ trunk/src/ta3d/src/misc/slist.h
@@ -271,9 +271,9 @@ namespace TA3D
 		}
 
 	private:
-		slist_node   *head;
-		slist_node   *last;
-		int          _size;
+		slist_node	*head;
+		slist_node	*last;
+		short		_size;
 
 	}; // class slist
